requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
telethon==1.33.1
pydantic
PyYAML==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dotenv==1.0.0